web_player/pyproject.toml
[project]
name = "music-player"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "cryptography>=46.0.1",
    "pyjwt>=2.10.1",
]
